Shifting offset now applied a time of calculation
Previously, the 'shift' value, which allows the xPPSOffset values to be
applied with an arbitrary shift with respect to the TIC time stamps, was
actually applied on import of the xPPSOffset data.  This patch applies
the shift between TIC and offset time stamps during xPPOffset
corrections.

Also, an option was added to turn on calculation and plotting of the
frequency depature.