Added Frac. Freq. Error and Histograms
Added calculation and plots of historgrams, so to speek.  I was going to
do histograms, but decided to do kernel density estimates.  I'd like to
do both on the same plot, but that isn't so easy.  I'll just leave KDE
for now.

Also, added calcuation of fractional frequncy error, i.e. first
differences.  Currently, it only shows in the preview, but I'll add AVAR
next.