Added GPS performance report (incomplete)
I added a call to sbfAnalyser that can run a predefined report.  This does not
work, though, as it is an X11 client application and requires an X server.  I
have decided to abandond that, for now, but it is possible to fix that later,
if I want, so I'm commiting it now anyway.