Added 3-day combined hybrid RINEX capability
mkRINEX+CGGTS.sh: new feature produces RINEX files that include a bit
of data from the previous day and the next day.  This was Nick's idea
and it produced improved continuity in the PPP results before we found
the smoothing option.  I've added it because I think we will use it
for in-house PPP, when we do that.