ptMon-pltDAQ.sh: changes for neutrino, bug fixes
Development occured on my home computer and most of the changes
in this commit are changes related to that.

Also, some bugs were fixed.  After plots are created in the tmpDir,
they should be coppied to ouputDir, but this was incorrect and was
fixed.  Secondly, the gnuplot Titles used '@' char, but this is a
special char, so I changed it to "at".