Added ptMon-all wrapper script for PT Monitoring
ptMon-all.sh: a wrapper script that runs all the PT Monitoring scripts.
Now that this is operational, I am expecting to make more of these
ploting scripts, and this script will call them.  This script is
intended to be run by cron and it also mangages STDERR from the plotting
scripts.  It puts errors into a log file.  The itention is that the logs
will also be processed by the script that builds the web pages.

I also changed the name of the DAQ output files.