High Load on neutrino + DAQ plot cycles
neutrino had a high load of >66 all day on Sunday.  I'm not sure if this is
normal or not.  In anycase, I don't want to exacerbate it, so I added a load
check to the DAQ plotting program.

I also refactored the DAQ plotting program to allow for different "cycles" or
plot groups/series.  The script checks a new parameter and can run different
sets of plots, rather than all the plots all the time.  This allows one to
run different sets of plots from ptMon-all or sparate cron jobs.

Also, ptMon-all accepts and passes on the "cycle" type.  It also runs both
sites' DAQ plots in parallel.