mkRIN: mtime control (sbfExtract() only)
If mkRINEX... ran immediately after download, it wouldn't need to
manipulate file timestamps.  But, currently, there is a back log, and
I've never autmated it.

And, it's only useful if you want to use timestamps to find SBF
extracted data.  But, I want to do this, now, to make the plotting tools
for the PT monitoring website.

So, In order to make modification times useful, I've added a mtime
manipulation code so that mkRIN preserves the modification time of the
SBF binary files *as* the mtime of the extracted data files (and their
compressed archive files in organizedData/).

In addition, this code adds a storage subroutine called by sbfExtract()
which contains the new logic.

But, this only works for sbfExtract() which is currently only used for
GLOTime extraction.  I think I will replace the other functions, since
doing so now has extra benefit.