pt-plotgen: Incompatible Change
A change introduced in the previous commit caused it to mailfunction
when 'call'ed (instead of loaded) and when passing complex plotting
commands (i.e. using 3:($2*10)) because $# is replaced when 'call'ed.
To fix this, I've implemented a backward compatible hook for the
usingCSV option and conerted the newer scripts to the 'load' method.

ptMon-rxClkBias: Also, this commit fixed a serious bug in the plotting
of rxClkBias data: the scale was not adjusted for nanoseconds.