mkRIN: Filled out MJD table to 2050
Also, I fixed logic error in the case that rin2cgg failed.  I wasn't able
to determine to cause of this on Jan 30 2014/1 Feb 2014, which is a good
test case for later, if it needs to get solved.  But, in such cases, unsafe
expansion of unilitialized variables is attempted, causing errors.
Also, changed trap to just say EXIT. not a big deal.