ptMon-fullCorr: alpha version
ptMon-fullCorr.sh is the final script for the PT Monitoring web-site.
It's purpose is to make graphs of the true PT-OT differences, including
all delays and Post-processing corrections.

This commmit only achives a small portion of that.  It extracts the
PPP data and plots it, alone.  No calculations are applied.  It is an
intermediary step.

While this version is tested and seems to work, it is not integrated
into the ptMon-all.sh script.