mkRIN: Stop pull/store RINEX unless doRIN processing enabled
There was (and still may be) a logic error in the complex way RINEX files
are processed.  The processing is separate from the retrieval/pull of old
RINEX data from the data store (organizedData/rinex) and the storing of
RINEX data to the same location.  The pull & store were still happening
even when RINEX processing was disabled (--norin or doRIN=no).  I turned
this off by added checkes at the pull and store points.  I cannot remember
why this was left going, but I think it used to serve some purpose.  We'll
see.