fullCorr: Kludge to force plot precedence
Added code to force the order of PPP result precedence.  This way, the last
source type in the PPP source lists, which is currently EMR-final gets
plotted on top of everything else.

Also, added code to reject all but final PPP data in some plots, like the
Level3/long-range plot.  This code is tested, but I turned if off again
thinking it would be nice to see the improvement of final results over
rapid results in the PT-Monitor web-site.