mkRIN: Changed find to look at change time
Switched from -newer to -cnewer so that TOKA SBF files, which are updated
twice a day, would get regenerated in the current automation system, which
picks files touched in the previous day.  This works, TOKA files get updated
so late that the time ranges would need to be expanded, causing rebuilds
of PT00 and PT01 data that do not need to be updated.  Either that, or
run a special reuild of TOKA data at the approriate time.

This may ultimately not work, and the alternative above will have to do.  But,
it may work and would be easiest to manage.